refactor test
